+++ ProjectScion/GameState.h
@@ -31,7 +31,7 @@ private:
 public:
     void Initialize(ScionEngine* game);
     void HandleInput(sf::RenderWindow* window);
-    void Update(double delta, bool isGameActive, bool isCoveredByOtherState);
+    void Update(float delta, bool isGameActive, bool isCoveredByOtherState);
     void Draw(sf::RenderWindow * window);
 };
