+++ ProjectScion/SplashScreenState.h
@@ -17,7 +17,7 @@ private:
 public:
     SplashScreenState(float duration, sf::Texture* image, sf::SoundBuffer* soundBuffer, bool clickToSkip = false);
     virtual void Initialize(ScionEngine* game);
-    virtual void Update(double delta, bool isGameActive, bool isCoveredByOtherState);
+    virtual void Update(float delta, bool isGameActive, bool isCoveredByOtherState);
     virtual void Draw(sf::RenderWindow* window);
     virtual void HandleInput(sf::RenderWindow* window);
 };
