+++ ProjectScion/SpriteAttribute.cpp
@@ -2,10 +2,10 @@
 
 SpriteAttribute::SpriteAttribute(int x, int y, sf::Texture& texture) : _sprite(texture)
 {
-    _sprite.setOrigin(x,y);
+    _sprite.setOrigin((float)x,(float)y);
 }
 
 SpriteAttribute::SpriteAttribute(int x, int y, sf::IntRect rect, sf::Texture& texture) : _sprite(texture, rect)
 {
-    _sprite.setOrigin(x,y);
+    _sprite.setOrigin((float)x,(float)y);
 }
